i2chipd.cpp define chip and file objects arrays for easier handling
Similar chip types can be processed in 'for loop' if the chip objects and
file objects are defined in array. The objects are ordered by their
I2C address in these arrays. For example TMP102 with addresses 0x48 - 0x4B
has array with length of four.

Some of the configuration settings can be changed by modifying constant
arrays at the beginning of the source code file.